added the canvas to html file to generate the picture later added variable "canvasColor" to header.component.ts to use it as value for canvas background added two function to change the color of canvas background ( switchColorWhite() and switchColorBlue() )
