--- src/header/header.component.html
@@ -1,46 +1,0 @@
-<style>
-  /* Add a black background color to the top navigation */
-  .topnav {
-    background: deepskyblue;
-    overflow: hidden;
-
-
-
-  }
-
-  /* Style the links inside the navigation bar */
-  .topnav a {
-    float: left;
-    text-align: center;
-    padding: 20px 20px;
-    text-decoration: none;
-    font-size: 17px;
-  }
-
-  /* Change the color of links on hover */
-  .topnav a:hover {
-    padding-bottom: 5px;
-    border-bottom-style: solid;
-    border-bottom-width: 2px;
-    width: fit-content;
-  }
-
-  /* Add a color to the active/current link */
-
-  .center {
-    margin: auto;
-    width: 20%;
-  }
-
-</style>
-<div class="topnav" [ngStyle]="{backgroundColor : color}">
-    <div class="center">
-      <a class="btn success"  [ngStyle]="{ color : 'white' }" (click)="switchColor()"  >White</a>
-
-
-      <a class="btn success" [ngStyle]="{ color : 'white' }" >Blue</a>
-
-    </div>
-
-
-</div>
